titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/eposs/digitization-documents-search/data-access-documents-search/src/lib/+state/documents-search.selectors.ts
import { createSelector } from '@ngrx/store';

import { selectDocumentSearchState } from './documents-search.reducer';

/**
 * The selectors for Data-Upload store
 */
const selectInvoiceListResponse = createSelector(
  selectDocumentSearchState,
  state => state.invoiceListResponse
);

const selectIsLoading = createSelector(
  selectDocumentSearchState,
  state => state.isLoading
);

const selectHasError = createSelector(
  selectDocumentSearchState,
  state => state.hasError
);

export const DocumentsSearchSelectors = {
  selectInvoiceListResponse,
  selectIsLoading,
  selectHasError
};
